ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
openai==1.51.2
whisper-timestamped==1.15.7
numpy==2.1.1
scikit-learn==1.5.2
rapidfuzz==3.9.7
python-dotenv==1.0.1
google-generativeai==0.7.2
requests==2.31.0
